requirements.txt
Flask==3.0.2
pyyaml==6.0.1
PyGithub==2.2.0
prometheus-flask-exporter==0.23.0